4.32 - add HHType to WHERE clause for insert to sys_Enrollment
Filter insert of inactive enrollments potentially relevant to SystemPath to sys_Enrollment by requiring a match on HHType as determined by the subquery (as of EntryDate) to tmp_Household.HHType (as of later of EntryDate/ReportStart).  No change to final output, but there's no point in inserting them and not counting them later because HHType doesn't match.