4.69 and 4.71 DQ for 3.917 fields
Correct identification of circumstances under which DateToStreetESSH, TimesHomelessPastThreeYears, and MonthsHomelessPastThreeYears are required for collection.  Note that THE SAMPLE CODE FOR THESE THREE FIELDS IS NOT CONSISTENT WITH THE SPECS because the specs are wrong about when they are required.  For FY2018, the HDX will ignore DQ values for these fields.