Update 9_1 to 9_3 LSAReport DQ and ReportDate.sql
Section 9.1 Change check of leftmost character in SSN from ">= '9'" to "= '9'" (issue #172) and add parentheses to WHERE clause (issue #173)

Section 9.2 Add check for Status1 is not null to criteria for DisablingCond1 (issue #174) and correction to criteria for Destination1 and Destination2.

Note that in practice, there is no need to check for non-NULL values for Status3 in the code because there are no non-NULL values in the column.