See Description - No Change to Output
According to SQL Server, the code to set EST/RRH/PSHGeography in section 6.5 was 75% of the actual cost of executing all of the code in this file.  I modified it and SQL Server now says it's 3% of the actual cost.  There is a minimal improvement in performance, but it's not as dramatic as I had hoped.